fix(combattracker.lic): updates
moved parser to it's own file

adjusted some of the data we were pushing.

no need to push creature to attack and flare tables when we already have a creature instance table we're linking to for that information.

no need to push crit info into the flare and attack event tables. We should utilize the damage components table for damage and crit type, location, rank and link back to the flare and attack tables.