fix(combattracker.lic): updates
made some adjustments to make compatible with assaults.

added some preliminary reports to help with data validation.

CombatTracker::Reporter.attack_report(attack, seq)
CombatTracker::Reporter.creature_report(noun, seq)

Can leave both options blank for a list of nouns. leaving seq blank or using -1 will show the last creature of that noun recorded. Can cycle through the seq for all of the mobs with that noun.

It creates the creature and attack index lists when ;reporter is ran. So for now run ;reporter then do your queries. Only need to update it after a hunt. This will be cleaned up when we incorporate real commands to call the reports.